cuad_qa/requirements.txt
accelerate==1.7.0
aioboto3==14.3.0
aiobotocore==2.22.0
aiodns==3.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
astor==0.8.1
attrs==24.3.0
awscli==1.38.3
backoff==2.2.1
bcrypt==4.3.0
black==24.10.0
blake3==1.0.4
boto3==1.37.3
botocore==1.37.3
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
compressed-tensors==0.9.3
contourpy==1.3.2
cryptography==44.0.3
-e git+ssh://git@github.com/marutilai/cuad-reinforcement-learning.git@d5a6637e3b5b1751a2a3266c0a9d84e4aedbb566#egg=cuad_qa
cycler==0.12.1
datasets==2.21.0
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docutils==0.16
einops==0.8.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flake8==7.2.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2024.6.1
gguf==0.16.3
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
hf-xet==1.1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
importlib_metadata==8.0.0
iniconfig==2.1.0
inquirerpy==0.3.4
interegular==0.3.3
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain-core==0.3.60
langsmith==0.3.42
lark==1.2.2
litellm==1.69.3
llguidance==0.7.20
llvmlite==0.44.0
lm-format-enforcer==0.10.11
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
mistral_common==1.5.4
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==2.2.5
omegaconf==2.4.0.dev3
openai==1.74.0
opencv-python-headless==4.11.0.86
openpipe==4.50.0
openpipe-art==0.3.7
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.8
orjson==3.10.18
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
panza==0.1.0
paramiko==3.5.1
partial-json-parser==0.2.1.1.post5
pathspec==0.12.1
peft==0.15.2
pendulum==3.1.0
pfzy==0.3.4
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
polars==1.29.0
prettytable==3.16.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
PuLP==3.1.1
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pycares==4.8.0
pycodestyle==2.13.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.46.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.6
rpds-py==0.25.0
rsa==4.7.2
runpod==1.7.9
s3transfer==0.11.3
safetensors==0.5.3
scipy==1.15.3
seaborn==0.13.2
semver==3.0.4
sentencepiece==0.2.0
sentry-sdk==2.28.0
setproctitle==1.3.6
setuptools==80.7.1
shellingham==1.5.4
six==1.17.0
skypilot==0.9.2
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tblib==3.1.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
torchao==0.11.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
tqdm-loggable==0.2
transformers==4.51.3
trl==0.15.2
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vllm==0.8.5.post1
wandb==0.19.11
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
